# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: VirtuProfs
basePath: /api
host: localhost:5100
schemes:
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  passport:
    type: apiKey
    in: header
    name: Authorization
  admin:
    type: apiKey
    in: header
    name: Authorization
  teacher:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /subject:
    x-swagger-router-controller: SubjectController
    post:
      operationId: setSubject
      summary: Set subject
      tags:
        - subject
      security:
        - admin: []
      parameters:
        - in: body
          name: subject
          description: Subject set
          schema:
            type: object
            required:
              - name
              - icon
              - callbackCost
              - questionCost
              - accessByGrade
            properties:
              _id:
                type: string
              name:
                type: object
                required:
                  - en
                  - fr
                properties:
                  en:
                    type: string
                  fr:
                    type: string
              description:
                type: string
              icon:
                type: string
              cost:
                type: number
                minimum: 0
              callbackCost:
                type: number
              questionCost:
                type: number
              accessByGrade:
                type: array
                items:
                  type: string
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
    delete:
      operationId: deleteSubjects
      summary: Delete subjects
      security:
        - admin: []
      tags:
        - subject
      parameters:
        - name: ids
          in: query
          type: array
          collectionFormat: csv
          items:
            type: string
          description: Delete subjects by ID
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
  /lessons/manage:
    x-swagger-router-controller: LessonController
    get:
      operationId: manageLessons
      summary: Get all own lessons
      security:
        - passport: []
      tags:
        - lesson
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
...
  /lesson:
    x-swagger-router-controller: LessonController
    get:
      operationId: getLessons
      summary: Get all lessons
      security:
        - admin: []
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
    post:
      operationId: setLesson
      summary: Set lesson
      security:
        - teacher: []
      tags:
        - lesson
      parameters:
        - in: body
          name: lesson
          description: lesson set
          schema:
            type: object
            required:
              - grade
              - subject
              - title
              - subtitle
            properties:
              _id:
                type: string
              grade:
                type: array
                items:
                  type: string
              subject:
                type: string
              title:
                type: string
              subtitle:
                type: string
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
    delete:
      operationId: deleteLessons
      summary: Delete lessons
      security:
        - teacher: []
      tags:
        - lesson
      parameters:
        - name: ids
          in: query
          type: array
          collectionFormat: csv
          items:
            type: string
          description: Delete lessons by ID
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
...
  /lesson/statistic:
    x-swagger-router-controller: LessonController
    get:
      operationId: lessonStatistic
      security:
        - passport: []
      summary: Get lesson statistic
      tags:
        - lesson
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
  /lesson/{id}:
    x-swagger-router-controller: LessonController
    get:
      operationId: getLesson
      summary: Get a lesson by ID
      parameters:
        - in: path
          name: id
          description: ID of lesson
          required: true
          type: string
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
  /lessonbysubj:
    x-swagger-router-controller: LessonController
    get:
      operationId: getLessonsBySubject
      summary: Get lessons by subject
      security:
        - passport: []
      tags:
        - lesson
      parameters:
        - name: subject
          in: query
          type: string
          description: Get quiz
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
...
definitions:
  Response:
    type: object
    properties:
      success:
        type: boolean
      errorCode:
        type: string
      errorStack:
        type: array
        items:
            type: object
      message:
        type: string
      data:
        type: object
    required:
      - success
...